Change the hlgroup of xolox#misc#msg#info
It used the hlgroup 'Title', but this does not accurately represent an informational message. As one can see from `:help hl-Title` and the example commands it lists, 'Title' is meant for section headings in long output, not a single-line message.

This inapplicability was especially noticeable in my color scheme (Solarized), which shows 'Title' in an orange-red color. This color does a good job of marking titles in long output, but it makes a single-line message look like an error.

Instead, I have made #info use no highlighting – the hlgroup 'None'. I think this hlgroup most accurately represents an “informational message”. It is the same hlgroup that searches are echoed with when pressing `n` and `N`, which also seem like informational messages.

While I am editing hlgroups, I also made the capitalization of their names the same as what Vim uses when you run the `:highlight` command. This makes multi-word names a little easier to read, and aids debugging against the output of `:highlight`.